sched: Use an integer for thread ids
Linux uses a 32-bit integer for pid_t, so let's do it as well. This is because
there are function in which we have to return our id back to the application.
One application is gettid, that we already have in the tree.

Theoretically, we could come up with a mapping between our 64-bit id and the
Linux one, but since we have to maintain the mapping anyway, we might as well
just use the Linux pids as our default IDs. The max size for that is 32-bit. It
is not enough if we're just allocating pids by bumping the counter, but again,
since we will have to maintain the bitmaps, 32-bit will allow us as much as 4
billion PIDs.

avi: remove unneeded #include

+++ include/sched.hh
@@ -371,7 +371,7 @@ private:
     cpu* _cpu;
     arch_thread _arch;
     arch_fpu _fpu;
-    unsigned long _id;
+    unsigned int _id;
     std::function<void ()> _cleanup;
     // When _ref_counter reaches 0, the thread can be deleted.
     // Starts with 1, decremented by complete() and also temporarily modified
@@ -397,7 +397,7 @@ public:
     // see cpu class
     lockless_queue_link<thread> _wakeup_link;
     static unsigned long _s_idgen;
-    static thread *find_by_id(unsigned long id);
+    static thread *find_by_id(unsigned int id);
 private:
     class reaper;
     friend class reaper;
